prepare files for merge

+++ src/main/java/ita/softserve/course_evaluation/repository/CourseRepository.java
@@ -2,13 +2,12 @@ package ita.softserve.course_evaluation.repository;
 
 import ita.softserve.course_evaluation.entity.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface CourseRepository extends JpaRepository<Course, Integer> {
+public interface CourseRepository extends JpaRepository<Course, Long> {
 
     Optional<Course> findByCourseName(String courseName);
 }
